Adjusted to PR change requests
